create package dto, update entity

+++ src/main/java/vn/com/shop/entity/CategoryEntity.java
@@ -12,7 +12,7 @@ public class CategoryEntity extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "CategoryID")
-    private int categoryID;
+    private Long categoryID;
 
     private String categoryName;
 }
